requirements.txt
keras==2.13.1
matplotlib==3.7.2
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
Requests==2.31.0
streamlit==1.28.1
streamlit_lottie==0.0.5
tensorflow==2.13.0

